Rework Wire Protocol
- Simplification of functions.
- Rework process state machine.
- Rework debugger functions changing flags to improve responsiveness.
- Rework WP receiver thread on all platforms to make it more responsive.
- Rework HAL interface on all platforms to make it more robust.

+++ src/CLR/WireProtocol/WireProtocol_HAL_Interface.c
@@ -16,7 +16,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 // provided as weak to be replaced by actual implementation by HAL interface
-__nfweak int WP_ReceiveBytes(uint8_t *ptr, unsigned short *size)
+__nfweak uint8_t WP_ReceiveBytes(uint8_t *ptr, uint32_t *size)
 {
     (void)(ptr);
     (void)(size);
@@ -26,7 +26,7 @@ __nfweak int WP_ReceiveBytes(uint8_t *ptr, unsigned short *size)
 }
 
 // provided as weak to be replaced by actual implementation by HAL interface
-__nfweak int WP_TransmitMessage(WP_Message *message)
+__nfweak uint8_t WP_TransmitMessage(WP_Message *message)
 {
     (void)(message);
